refactor(display): extract display configuration building logic
